Remove existing package before installing.
Ideally, one would want yum to upgrade the existing package using the
fresh, built one,. This does not woek, though. rpm can do it using
rpm -U, but that means problems wit deps.

This means that if this operation fails somehow, the net effect might
be to destroy existing package without getting a new in place.